Why are we casting integers as integers?
I honestly don’t know. I’m not sure if this is something that I don’t
see the value in due to a lack of experience, or if it is in fact
totally pointless and insane. For the sake of preserving my sanity, I’m
removing it to see what happens. WILL THE WORLD BURN?! TIME WILL TELL