Refactored & renamed navigation template tags
The navigation template tags are now closer to what’s in WordPress
core. I would have liked to use the core tag for single post navigation
here, but it lacks some of the nicer features of the old template tags.
It’s also a wrapper for wrapper functions, which seems rather odd to
me… In any event, this preserves the features of the older template
tags in a single handy function and adds some additional filtering
capabilities.